+++ cifar10_models/__init__.py
@@ -5,4 +5,4 @@ from .densenet import *
 from .resnet_orig import *
 from .googlenet import *
 from .inception import *
-from .mlp import *
+from .simple import *
